Some minor style changes
